Simplify connecting to Kafka by using sarama-heroku
This change has two main advantages: it reduces the amount of code
needed to connect to Heroku Kafka and it improves security around
certificate verification.

The demo code includes a lot of boilerplate for verifying certificates
that would likely be copy and pasted into apps using this code as a
starting point. It's easy to imagine someone copying the code but
leaving out the verification step because they don't understand it.

Additionally, the current method is fundamentally flawed in that it only
verifies certificates on boot. The TLSConfig created by sarama-heroku
includes a custom certificate verifier that will validate the server's
certificate against the trusted cert on each connect. This would allow
validating any new servers that appear in the metadata after launch
without a restart.

Finally, this makes integrating with Heroku Kafka a lot easier and
consolidates the tricky to understand certificate process into a
reusable component. I was able to remove 125 lines from main.go,
allowing it to be more focused on using Kafka and not connecting.